Add LfStaticVarCompensator
